+++ src/workbench/index.ts
@@ -9,24 +9,24 @@ export function getWorkbench(configuration: Configuration, variant: string) {
   if (variant === "dark") {
     switch (configuration.darkWorkbench) {
       case "material":
-        return materialWorkbench(palette);
+        return materialWorkbench(palette, configuration, "dark");
       case "flat":
-        return flatWorkbench(palette);
+        return flatWorkbench(palette, configuration, "dark");
       case "high-contrast":
-        return highContrastWorkbench(palette);
+        return highContrastWorkbench(palette, configuration, "dark");
       default:
-        return materialWorkbench(palette);
+        return materialWorkbench(palette, configuration, "dark");
     }
   } else {
     switch (configuration.lightWorkbench) {
       case "material":
-        return materialWorkbench(palette);
+        return materialWorkbench(palette, configuration, "light");
       case "flat":
-        return flatWorkbench(palette);
+        return flatWorkbench(palette, configuration, "light");
       case "high-contrast":
-        return highContrastWorkbench(palette);
+        return highContrastWorkbench(palette, configuration, "light");
       default:
-        return materialWorkbench(palette);
+        return materialWorkbench(palette, configuration, "light");
     }
   }
 }